+++ src/utils.c
@@ -56,36 +56,36 @@ double get_time_in_seconds() {
     return ts.tv_sec + ts.tv_nsec / 1e9;
 }
 
-void benchmark_function(void (*func)(float**, int, double*), float** matrix, int n, const char* func_name) {
+void benchmark_function(void (*func)(float**, int, long double*), float** matrix, int n, const char* func_name) {
     double total_time = 0.0;
     for (int i = 0; i < NUM_RUNS; i++) {
-        double time;
+        long double time;
         func(matrix, n, &time);
         total_time += time;
     }
     printf("%s average time: %f seconds\n", func_name, total_time / NUM_RUNS);
 }
 
-void is_symmetric_wrapper(float** matrix, int n, double* time) {
+void is_symmetric_wrapper(float** matrix, int n, long double* time) {
     is_symmetric(matrix, n, time);
 }
 
-void is_symmetric_implicit_wrapper(float** matrix, int n, double* time) {
+void is_symmetric_implicit_wrapper(float** matrix, int n, long double* time) {
     is_symmetric_implicit(matrix, n, time);
 }
 
-void is_symmetric_omp_wrapper(float** matrix, int n, double* time) {
+void is_symmetric_omp_wrapper(float** matrix, int n, long double* time) {
     is_symmetric_omp(matrix, n, time);
 }
 
-void transpose_wrapper(float** matrix, int n, double* time) {
+void transpose_wrapper(float** matrix, int n, long double* time) {
     transpose(matrix, n, time);
 }
 
-void transpose_omp_wrapper(float** matrix, int n, double* time) {
+void transpose_omp_wrapper(float** matrix, int n, long double* time) {
     transpose_omp(matrix, n, time);
 }
 
-void transpose_omp_block_based_wrapper(float** matrix, int n, double* time) {
+void transpose_omp_block_based_wrapper(float** matrix, int n, long double* time) {
     transpose_omp_block_based(matrix, n, BLOCK_SIZE, time); // Assuming block size of 64
 }
